+++ NEWS.md
@@ -16,6 +16,7 @@
 * Implement `facet_strata()` to create a random group of size `n_strata` to put the data into (#32). Add support for `along`, and `fun`.
 * Implement `facet_sample()` to create facetted plots with a set number of keys inside each facet. (#32). 
 * `add_` functions now return a `tsibble()` (#49).
+* Fixed bug where `stratify_keys()` didn't assign an equal number of keys per strata (#55)
 
 # brolgar 0.0.1.9000